refactor: move the function trim() to a separate file
