+++ func.js
@@ -1,0 +1,15 @@
+function trim(x, characters=" \t\w") {
+  var start = 0;
+  while (characters.indexOf(x[start]) >= 0) {
+    start += 1;
+  }
+  var end = x.length - 1;
+  while (characters.indexOf(x[end]) >= 0) {
+    end -= 1;
+  }
+  return x.substr(start, end - start + 1);
+}
+
+module.exports = {
+  trim,
+}
